profrec/requirements.txt
Flask
requests
psycopg2-binary
beautifulsoup4
RateMyProfessorAPI
python-dotenv
gunicorn
Flask-Cors
aiohttp